+++ test_apps/lowest-peers-stimulus3/assets/app.js
@@ -1,0 +1,20 @@
+/*
+ * This file is part of the Symfony Webpack Encore package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+'use strict';
+
+import { startStimulusApp } from '@symfony/stimulus-bridge';
+
+const app = startStimulusApp(require.context(
+    '@symfony/stimulus-bridge/lazy-controller-loader!./controllers',
+    true,
+    /\.[jt]sx?$/,
+));
+
+console.log('Stimulus app started', app);