+++ test_apps/lowest-peers-svelte3/assets/app.js
@@ -1,0 +1,17 @@
+/*
+ * This file is part of the Symfony Webpack Encore package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+'use strict';
+
+import App from './App.svelte';
+
+const target = document.getElementById('app');
+if (target) {
+    new App({ target });
+}