+++ test_apps/lowest-peers-typescript-fork8/webpack.config.js
@@ -1,0 +1,12 @@
+import Encore from '@symfony/webpack-encore';
+
+Encore
+    .setOutputPath('public/build/')
+    .setPublicPath('/build')
+    .enableSingleRuntimeChunk()
+    .enableTypeScriptLoader()
+    .enableForkedTypeScriptTypesChecking()
+    .addEntry('app', './assets/app.ts')
+;
+
+export default await Encore.getWebpackConfig();