+++ test_apps/lowest-peers-vue/assets/App.vue
@@ -1,0 +1,11 @@
+<template>
+    <h1>{{ msg }}</h1>
+</template>
+
+<script>
+export default {
+    data() {
+        return { msg: 'Hello world!' };
+    },
+};
+</script>