+++ test_apps/lowest-peers-vue/assets/app.js
@@ -1,0 +1,18 @@
+/*
+ * This file is part of the Symfony Webpack Encore package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+'use strict';
+
+import { createApp } from 'vue';
+import App from './App.vue';
+
+const el = document.getElementById('app');
+if (el) {
+    createApp(App).mount(el);
+}